move the act plumbing out, so act-tools fits under the backstop again
#423 was written against `release/v2.12.0`, where `act-tools.ts` had room. On
`main`, 48 commits further along, the same addition takes it to 1003 lines and
`max-lines` refuses it. The cap is the backstop for cohesion, so the answer is
a seam rather than a bigger number.

`actCommand` and `resolveActTarget` are the part that is identical across all
three acting tools — dispatch an action to the bridge, and turn a `target`
into a ref before the action window opens. The tools are three different units
of work and stay together; what they have in common moved.

No behaviour change: the functions are unmoved, only re-homed and exported.
